+++ lib/resque-exceptional.rb
@@ -1,8 +1,4 @@
-require 'zlib'
-require 'net/http'
-require 'net/https'
-require 'time'
-require 'digest/md5'
 require 'resque'
+require 'honeybadger'
 
-require 'resque/failure/exceptional'
+require 'resque/failure/honeybadger'